# SWAGGER_OPENAPI/datavault-broker-openapi.yaml
openapi: 3.1.0
info:
  title: DataVault Broker API
  description: |-
    API documentation for the DataVault Broker module.
    All endpoints require authentication via the X-UserID and X-Client-Key headers.
  version: 0.0.0 (4b97dc0)
  x-generated-at: 2026-03-12T10:56:37.585546
servers:
- url: http://localhost
  description: Generated server url
security:
- X-UserID-Header: []
- X-Client-Key-Header: []
...
paths:
...
  /admin/archivestores:
    get:
      tags:
      - admin-archive-store-controller
      operationId: getArchiveStores
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchiveStore"
    put:
      tags:
      - admin-archive-store-controller
      operationId: editArchiveStore
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveStore"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveStore"
    post:
      tags:
      - admin-archive-store-controller
      operationId: addArchiveStore
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveStore"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveStore"
...
  /vaults/retrievequeuecount:
    get:
      tags:
      - statistics-controller
      summary: Get the count of Retrieves in queue
      description: Retrieves the number of Retrieves currently in the queue.
      operationId: getRetrievesQueuedCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
  /vaults/retrieveinprogress:
    get:
      tags:
      - statistics-controller
      summary: Get Retrieves in progress
      description: Retrieves a list of Retrieves currently in progress.
      operationId: getRetrievesInProgress
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Retrieve"
  /vaults/retentionpolicycount/{status}:
    get:
      tags:
      - statistics-controller
      summary: Get count of Vaults by Retention Policy Status
      description: Retrieves the number of Vaults matching a specific Retention Policy
        status.
      operationId: getPolicyStatusCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: status
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /vaults/depositqueuecount:
    get:
      tags:
      - statistics-controller
      summary: Get the count of Deposits in queue
      description: Retrieves the number of Deposits currently in the queue.
      operationId: getDepositsQueueCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
  /vaults/depositinprogress:
    get:
      tags:
      - statistics-controller
      summary: Get Deposits in progress
      description: Retrieves a list of Deposits currently in progress.
      operationId: getDepositsInProgress
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
...
  /statistics/size:
    get:
      tags:
      - statistics-controller
      summary: Get the total size of Vaults
      description: Retrieves the combined size of all Vaults in the system.
      operationId: getVaultsSize
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /statistics/retrieveinprogresscount:
    get:
      tags:
      - statistics-controller
      summary: Get the count of Retrieves in progress
      description: Retrieves the number of Retrieves currently in progress.
      operationId: getRetrievesInProgressCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
  /statistics/retrievecount:
    get:
      tags:
      - statistics-controller
      summary: Get the total count of Retrieves
      description: Retrieves the total number of Retrieves in the system.
      operationId: getRetrievesCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /statistics/eventcount:
    get:
      tags:
      - statistics-controller
      summary: Get the total count of Events
      description: Retrieves the total number of Events in the system.
      operationId: getEventCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /statistics/depositinprogresscount:
    get:
      tags:
      - statistics-controller
      summary: Get the count of Deposits in progress
      description: Retrieves the number of Deposits currently in progress.
      operationId: getDepositsInProgressCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
  /statistics/depositcount:
    get:
      tags:
      - statistics-controller
      summary: Get the total count of Deposits
      description: Retrieves the total number of Deposits in the system.
      operationId: getDepositsCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
  /statistics/count:
    get:
      tags:
      - statistics-controller
      summary: Get the total count of Vaults
      description: Retrieves the total number of Vaults in the system.
      operationId: getVaultsCount
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /admin/retrieves/count:
    get:
      tags:
      - admin-controller
      summary: Get the total count of Retrieves
      description: "Retrieves the total number of Retrieves in the system, with an\
        \ optional query filter."
      operationId: getRetrievesCount_1
      parameters:
      - name: X-UserID
        in: header
        description: DataVault Broker User ID
        required: true
        schema:
          type: string
      - name: query
        in: query
        description: Retrieve query field
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /admin/deposits/count:
    get:
      tags:
      - admin-controller
      summary: Get the total count of Deposits
      description: "Retrieves the total number of Deposits in the system, with an\
        \ optional query filter."
      operationId: getDepositsCount_1
      parameters:
      - name: X-UserID
        in: header
        description: DataVault Broker User ID
        required: true
        schema:
          type: string
      - name: query
        in: query
        description: Deposit query field
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
...
  /admin/archivestores/{archiveStoreId}:
    get:
      tags:
      - admin-archive-store-controller
      operationId: getArchiveStore
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: archiveStoreId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveStore"
    delete:
      tags:
      - admin-archive-store-controller
      operationId: deleteArchiveStore
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: archiveStoreId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    DataCreator:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        name:
          type: string
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        visible:
          type: boolean
        crisId:
          type: string
    RetentionPolicy:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        url:
          type: string
        minRetentionPeriod:
          type: integer
          format: int32
        extendUponRetrieval:
          type: boolean
        inEffectDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dataGuidanceReviewed:
          type: string
          format: date
        engine:
          type: string
          deprecated: true
        sort:
          type: integer
          format: int32
          deprecated: true
        minDataRetentionPeriod:
          type: string
          deprecated: true
        policyInfo:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        password:
          type: string
        email:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
    Vault:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        creationTime:
          type: string
          format: date-time
        grantEndDate:
          type: string
          format: date
        reviewDate:
          type: string
          format: date
        name:
          type: string
        description:
          type: string
        retentionPolicy:
          $ref: "#/components/schemas/RetentionPolicy"
        retentionPolicyStatus:
          type: integer
          format: int32
        retentionPolicyExpiry:
          type: string
          format: date-time
        retentionPolicyLastChecked:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          deprecated: true
        dataset:
          $ref: "#/components/schemas/Dataset"
        projectId:
          type: string
        snapshot:
          type: string
        affirmed:
          type: boolean
        notes:
          type: string
        estimate:
          type: string
          enum:
          - UNDER_100GB
          - UNDER_10TB
          - OVER_10TB
          - UNKNOWN
        contact:
          type: string
        pureLink:
          type: boolean
        size:
          type: integer
          format: int64
        vaultReviewUnderway:
          type: boolean
        mostRecentVaultReview:
          $ref: "#/components/schemas/VaultReview"
        dataCreator:
          type: array
          items:
            $ref: "#/components/schemas/DataCreator"
          writeOnly: true
    VaultReview:
      type: object
      properties:
        id:
          type: string
        creationTime:
          type: string
          format: date-time
        vault:
          $ref: "#/components/schemas/Vault"
        newReviewDate:
          type: string
          format: date
          deprecated: true
          writeOnly: true
        oldReviewDate:
          type: string
          format: date
        actionedDate:
          type: string
          format: date-time
        comment:
          type: string
    Deposit:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit
        version:
          type: integer
          format: int64
        creationTime:
          type: string
          format: date-time
          description: Date that the vault was created
        vault:
          $ref: "#/components/schemas/Vault"
        depositPaths:
          type: array
          items:
            $ref: "#/components/schemas/DepositPath"
        depositChunks:
          type: array
          items:
            $ref: "#/components/schemas/DepositChunk"
        status:
          type: string
          description: Status of the Deposit
          enum:
          - NOT_STARTED
          - IN_PROGRESS
          - COMPLETE
          - FAILED
          - DELETE_IN_PROGRESS
          - DELETED
          - DELETE_FAILED
        name:
          type: string
          description: Deposit name to briefly describe the Deposit
        description:
          type: string
          description: Deposit description to provide more information about the Deposit
        hasPersonalData:
          type: boolean
          description: Whether the deposit contains personal data or not
        personalDataStatement:
          type: string
          description: Description of the nature of the personal data
        bagId:
          type: string
          description: ID of the bag associated with this Deposit
        archiveSize:
          type: integer
          format: int64
        archiveDigest:
          type: string
        archiveDigestAlgorithm:
          type: string
        numOfChunks:
          type: integer
          format: int32
        encArchiveDigest:
          type: string
        encIV:
          type: string
          format: byte
        fileOrigin:
          type: string
          description: Origin of the deposited filepath
        shortFilePath:
          type: string
          description: Short version of the origin of the deposited filepath
        filePath:
          type: string
          description: Filepath of the origin deposit
        user:
          $ref: "#/components/schemas/User"
        nonRestartJobId:
          type: string
        size:
          type: integer
          format: int64
    DepositChunk:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit Path
        chunkNum:
          type: integer
          format: int32
          description: Filepath of the origin deposit
        archiveDigest:
          type: string
        archiveDigestAlgorithm:
          type: string
        encIV:
          type: string
          format: byte
        ecnArchiveDigest:
          type: string
    DepositPath:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit Path
        filePath:
          type: string
          description: Filepath of the origin deposit
        pathType:
          type: string
          enum:
          - FILESTORE
          - USER_UPLOAD
...
    ArchiveStore:
      type: object
      properties:
        id:
          type: string
        storageClass:
          type: string
        label:
          type: string
        retrieveEnabled:
          type: boolean
        properties:
          type: object
          additionalProperties:
            type: string
...
    Retrieve:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        deposit:
          $ref: "#/components/schemas/Deposit"
        note:
          type: string
        status:
          type: string
          enum:
          - NOT_STARTED
          - IN_PROGRESS
          - COMPLETE
          - FAILED
        retrievePath:
          type: string
        hasExternalRecipients:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
...
  securitySchemes:
    X-UserID-Header:
      type: apiKey
      description: "User ID header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_USER_ID`."
      name: X-UserID
      in: header
    X-Client-Key-Header:
      type: apiKey
      description: "Client Key header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_CLIENT_KEY`."
      name: X-Client-Key
      in: header